requirements.txt
Flask==2.3.3
numpy==1.24.3
Flask-CORS==4.0.0
pandas==2.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
google-generativeai
